Proposition: Add link to open-source R textbook
I propose to add a link to the open source textbook "Tidy Finance with R" (print release with CRC Press in 2023) in which we provide code for "common" tasks such as CRSP/Compustat cleaning, portfolio sorts, portfolio optimization and much more. I hope the addition is of interest to the readers of the task view "Empirical Finance".